Add /opt/ros/rosdistro/lib to default library path.
Previously this library path was added to LD_LIBRARY_PATH by environment
hooks in other packages. However this means that pure-CMake packages
which don't depend on ament libraries and do not set environment hooks
of their own aren't found by dependents which are also pure cmake.

An example of this is fastrtps and fastcdr, both of which are
third-party cmake packages which do not register environment hooks.

This bug was previously masked for fastrtps and fastcdr but a
mis-applied dependency on ament_cmake rather than cmake in fastrtps.