Use GNUInstallDirs to determine multiarch library dir (#20)
The CMAKE_INSTALL_LIBDIR variable set by GNUInstallDirs contains the
platform's preferred library directory name. On most platforms, this
will be either 'lib' or 'lib64'. Debian does things a little
differently - it always uses 'lib', but creates a subdirectory for
each supported architecture.

In the end, the use case that drives this extra environment hook is a
package that uses CMAKE_INSTALL_LIBDIR instead of using the ament
pattern of just installing directly to 'lib'. Rather than predict the
debian-style pattern, it supports more platforms and is less fragile to
just let GNUInstallDirs do its thing and tell us where the libraries are
expected to be placed.

One curiosity I noticed is that during local builds, GNUInstallDirs
continued to report the 'lib' directory for use on Ubuntu. After looking
at the underlying script, it seems that because we're using a custom
CMAKE_INSTALL_PREFIX, the script doesn't suffix the arch-specific
directory like I was expecting. It turns out that debhelper actually
specifies a static CMAKE_INSTALL_LIBDIR value that overrides
GNUInstallDirs. At first I thought we'd need extra logic to ensure the
subdirectory is suffixed despite our non-standard CMAKE_INSTALL_PREFIX,
but we actually just want to just let the dance play out normally,
because that's what the downstream projects are going to do, too.

So in the end, we should see:
* Local builds on Debian should report 'lib', and this skip the hook
* Debian binary package builds should report 'lib/arch-triple', as
  debhelper will override the script
* Other platforms will report 'lib64' or 'lib' based on the platforms
  'void *' size

Signed-off-by: Scott K Logan <logans@cottsay.net>